Rename `imagenet_trainer.py` => `imagenet_trainer_tf.py`
