+++ tests/integration_tests/trainers/test_imagenet_trainer_tf.py
@@ -1,11 +1,11 @@
-"""Integration tests for trainers.imagenet_trainer"""
+"""Integration tests for trainers.imagenet_trainer_tf"""
 
 import tensorflow as tf
 
 from datasets.imagenet_dataset import ImageNetDataSet
 from datasets.tf_data_loader import TFDataLoader
 from networks.alexnet_tf import AlexNet
-from trainers.imagenet_trainer import ImageNetTrainer
+from trainers.imagenet_trainer_tf import ImageNetTrainer
 from utils.test_utils import df_images
